// src/interfaces/dataTest.interface.ts
export interface DataTest {
    data:                      number[];
    label:                     string;
/*     backgroundColor:           string;
    borderColor:               string;
    pointBackgroundColor:      string;
    pointBorderColor:          string;
    pointHoverBackgroundColor: string;
    pointHoverBorderColor:     string;
    fill:                      string; */
}